+++ app/domain/aggregates/song/entities/vocabulary.rb
@@ -27,7 +27,7 @@ module LyricLab
       def gen_filtered_words(text, openai_api_key)
         raise 'Language Level hasn\'t been defined yet' if language_level.nil?
 
-        openai_api = LyricLab::OpenAI.new(openai_api_key)
+        openai_api = LyricLab::OpenAI::API.new(openai_api_key)
         voc_factory = LyricLab::Vocabulary::VocabularyFactory.new(openai_api)
         self.filtered_words = voc_factory.create_filtered_words_from_text(text, language_level)
       end
